requirements.txt
einops==0.8.0
gymnasium[atari]==1.0.0a2
ale-py==0.9.1
hydra-core==1.3.2
opencv-python==4.7.0.72
protobuf==3.20.*
psutil==5.8.0
pygame>=2.1.2
tqdm>=4.62.3
wandb>=0.15.9
nltk==3.8.1
loguru
yet-another-retnet>=0.5.0
numpy==1.24.2
rliable
dm_control==1.0.19
dm-env==1.6
mujoco==3.1.5
mujoco-utils==0.0.6
gdown
